B2: Add option to hide files instead of deleting
This allows making proper use of B2's file versioning system and enables
restic to work with application keys that don't have the deleteFiles capability.

+++ internal/backend/b2/b2.go
@@ -253,7 +253,11 @@ func (be *b2Backend) Remove(ctx context.Context, h restic.Handle) error {
 	defer be.sem.ReleaseToken()
 
 	obj := be.bucket.Object(be.Filename(h))
-	return errors.Wrap(obj.Delete(ctx), "Delete")
+	if be.cfg.SoftDelete {
+		return errors.Wrap(obj.Hide(ctx), "Hide")
+	} else {
+		return errors.Wrap(obj.Delete(ctx), "Delete")
+	}
 }
 
 type semLocker struct {
